Planner: extract optimizer module; optimize() is pass-through
